Fix typing and run it in CI
We need the toml stubs package for typing to work.

Also typing wasn't runnin in CI so we should do that. Note that the reason we pass the `-e` parameter and don't just do `tox` is because we only want it to run for the currently set up python version.